Simplifying the 'elements' verification.
Hi Adam!

In the 18 line, your first verification is just to identify if 'elements' is not empty and in groovy you don't need to call the size() method to compare the result (two process). You can just call your map, because [:] is false. So... Something like [a:1] will return true.